not sure if I am doing the csv mapping properly, still erroring when I run it
