+++ DataImporter/Properties/launchSettings.json
@@ -1,0 +1,11 @@
+{
+  "profiles": {
+    "WSL": {
+      "commandName": "WSL2",
+      "distributionName": ""
+    },
+    "DataImporter": {
+      "commandName": "Project"
+    }
+  }
+}
