+++ GardenersMultitool.Api/CustomConfigurations/PlantCache.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using PlantDataImporter;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace GardenersMultitool.Api.CustomConfigurations
