+++ GardenersMultitool.Domain/Helpers/ZipcodeHardiness.cs
@@ -1,7 +1,6 @@
 ﻿using CSharpFunctionalExtensions;
 using GardenersMultitool.Domain.Entities;
 using GardenersMultitool.Domain.ValueObjects;
-using GardenersMultitool.Domain.ValueObjects.HabitationZone;
 using System;
 using System.Collections.Generic;
