+++ GardenersMultitool.sln
@@ -1,13 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31515.178
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GardenersMultitool.Domain", "GardenersMultitool.Domain\GardenersMultitool.Domain.csproj", "{8F33B501-108B-4DA7-A2BB-4D989BCF0284}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GardenersMultitool.Domain.Tests", "GardenersMultitool.Domain.Tests\GardenersMultitool.Domain.Tests.csproj", "{19C99C7A-CE4D-4746-A0FC-8E43784C156F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlantDataImporter", "PlantDataImporter\PlantDataImporter.csproj", "{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GardenersMultitool.Api", "GardenersMultitool.Api\GardenersMultitool.Api.csproj", "{FE23C7ED-C15E-4DA5-8031-72FE5769A870}"
 EndProject
@@ -18,7 +16,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{
 		docker-compose.yml = docker-compose.yml
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZipcodeDataImporter", "ZipcodeDataImporter\ZipcodeDataImporter.csproj", "{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{158E2DCF-04B2-41DE-8370-2EDD3D716D6D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataImporter", "DataImporter\DataImporter.csproj", "{96D734E2-C4D8-4054-B707-5C67D1F3952F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -54,18 +54,6 @@ Global
 		{19C99C7A-CE4D-4746-A0FC-8E43784C156F}.Release|iPhone.Build.0 = Release|Any CPU
 		{19C99C7A-CE4D-4746-A0FC-8E43784C156F}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{19C99C7A-CE4D-4746-A0FC-8E43784C156F}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Release|iPhone.Build.0 = Release|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{1D78F622-EC35-45AC-9F04-AC4FE4B8AC01}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 		{FE23C7ED-C15E-4DA5-8031-72FE5769A870}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FE23C7ED-C15E-4DA5-8031-72FE5769A870}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FE23C7ED-C15E-4DA5-8031-72FE5769A870}.Debug|iPhone.ActiveCfg = Debug|Any CPU
@@ -78,21 +66,24 @@ Global
 		{FE23C7ED-C15E-4DA5-8031-72FE5769A870}.Release|iPhone.Build.0 = Release|Any CPU
 		{FE23C7ED-C15E-4DA5-8031-72FE5769A870}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{FE23C7ED-C15E-4DA5-8031-72FE5769A870}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Release|iPhone.Build.0 = Release|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{B0DD8EF3-0617-4BC5-AEF3-45BB499075D9}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Release|iPhone.Build.0 = Release|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{96D734E2-C4D8-4054-B707-5C67D1F3952F} = {158E2DCF-04B2-41DE-8370-2EDD3D716D6D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9F1032B5-3365-40C6-ACD3-2A11728C35F3}
