--- PlantDataImporter/Properties/launchSettings.json
@@ -1,8 +1,0 @@
-{
-  "profiles": {
-    "PlantDataImporter": {
-      "commandName": "Project",
-      "commandLineArgs": "\"Permaculture_Plant_CSVs\""
-    }
-  }
-}
