--- ZipcodeDataImporter/ZipcodeDataImporter.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GardenersMultitool.Domain\GardenersMultitool.Domain.csproj" />
-    <ProjectReference Include="..\PlantDataImporter\PlantDataImporter.csproj" />
-  </ItemGroup>
-
-</Project>
